+++ packages/zowe-explorer/__tests__/__theia__/theia/extension.theiaFirefox.ts
@@ -14,6 +14,7 @@ import { Builder, By, Key, until } from "selenium-webdriver";
 import * as firefox from "selenium-webdriver/firefox";
 import { TheiaLocator, DatasetsLocators, UssLocators, JobsLocators } from "./Locators";
 
+const SHORTSLEEPTIME = 2000;
 const WAITTIME = 30000;
 let driverFirefox: any;
 
@@ -72,6 +73,10 @@ export async function addProfileDetailsInUss(profileName: string) {
     const ussProfileName = await driverFirefox.findElement(By.xpath(UssLocators.emptyInputBoxXpath));
     ussProfileName.sendKeys(profileName);
     ussProfileName.sendKeys(Key.ENTER);
+    await driverFirefox.sleep(SHORTSLEEPTIME);
+    const addToAllTrees = await driverFirefox.findElement(By.xpath(JobsLocators.emptyInputBoxXpath));
+    addToAllTrees.sendKeys("No");
+    addToAllTrees.sendKeys(Key.ENTER);
 }
 
 export async function addProfileDetailsInJobs(profileName: string) {